fix(bganim): defined float-to-integer conversion in fastCosRad and ember
Both sites convert a float built from tMs, which reaches 4.3e9 before it
wraps, straight to a narrower integer type. Once the scaled value passes
the target type's range the conversion is undefined rather than wrapping,
so the & 255 in fastCosRad was masking a value the conversion was never
required to produce, and ember's virtual-time word was undefined at the
top of the speed range.

fastCosRad is the shared helper, so this affected every animation that
calls it: lava, caustics, mandala, silk, fireflies and orbits. All call
sites are in frame() rather than a per-pixel loop, a few hundred calls per
frame at most, so widening the intermediate to int64_t costs nothing
measurable. Noted that in the header so it does not get copied into a
band() loop later.

Found by adding -fsanitize=float-cast-overflow to the sweep. GCC does not
fold it into -fsanitize=undefined, so the harness had been blind to this
class from the start. With it enabled the whole fleet is clean: 13
animations, 49-51 parameter sets each, six times per set, every band.

Golden frames and timings unchanged; ember stays bit-exact.

+++ src/display/ui/default/bganim/AnimEmber.cpp
@@ -139,7 +139,14 @@ void frame(uint32_t tMs, int, int, const uint8_t p[4]) {
     const float spd = speedMul(p[0]);
     // Speed scales virtual time; a param change causes one phase jump, which
     // the slow breathing envelope absorbs invisibly.
-    const uint32_t vt = static_cast<uint32_t>(tMs * spd);
+    // Virtual time is deliberately modular (every use below is a shift-and-mask
+    // into a 1024-entry sine table), but tMs * spd reaches ~2.9e10 at the top
+    // of the speed range before tMs wraps, and converting a float that large
+    // straight to uint32_t is undefined rather than wrapping. Go through double
+    // (float's 24-bit mantissa cannot hold tMs near its wrap anyway) and then
+    // int64_t, where the conversion is defined, and let integer-to-unsigned do
+    // the modular reduction.
+    const uint32_t vt = static_cast<uint32_t>(static_cast<int64_t>(static_cast<double>(tMs) * spd));
     const float pulseGain = p[3] / 100.0f;
     const float s1 = sin1024((vt * STEP1) >> 22) * (1.0f / SIN_AMP);
     const float s2 = sin1024(((vt * STEP2) + PHOFF2) >> 22) * (1.0f / SIN_AMP);
